แก้ css

+++ src/components/calendar_new/roomlist.jsx
@@ -21,6 +21,8 @@ export const RoomList = ({ academicYearTermId } ) => {
 
     const handelDeleteStudingTime = async (classroomId) => {
         try {
+            const text = "คุณต้องการจะลบปฎิทินหรือไม่"
+            if(!confirm(text)) return;
             const response = await axios.delete(`${HOSTNAME}/a/studingtime/${classroomId}`);
             if(response.status === 200){
                 fecthClassrooms();
@@ -38,33 +40,48 @@ export const RoomList = ({ academicYearTermId } ) => {
 
     return(
         <div>
-        <div className="rounded-lg border border-gray-200 mb-2">
-            <div className="overflow-x-auto rounded-t-lg">
+        <div className="border shadow-md border-gray-200 mb-2">
+            <div className="overflow-x-auto">
                 <table className="min-w-full divide-y-2 divide-gray-200 bg-white text-sm">
                     <thead className="ltr:text-left rtl:text-right">
-                        <tr>
+                        <tr className="h-12 shadow-md">
                             <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-900">ระดับชั้น</th>
                             <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-900">ห้องเรียน</th>
-                            <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-900">ลบปฎิทิน</th>
-                            <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-900" colSpan={2}>ปฎิทิน</th>
+                            <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-900">ปฎิทิน</th>
+                            <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-900">จัดการ</th>
+                            
                         </tr>
                     </thead>
                     <tbody className="divide-y divide-gray-200">
                             {
                                 sliceHolidayList.length > 0 && sliceHolidayList.map((classroom) => (
-                                    <tr key={classroom.classId}>
+                                    <tr key={classroom.classId} className="even:bg-slate-100/70 text-center">
                                         <td className="whitespace-nowrap text-center px-4 py-2 text-gray-700">{classroom.classLevel}</td>
                                         <td className="whitespace-nowrap text-center px-4 py-2 text-gray-700">{classroom.classRoom}</td>
-                                        <td className="whitespace-nowrap text-center px-4 py-2 text-red-700 cursor-pointer" onClick={() => handelDeleteStudingTime(classroom.classId)}>ลบปฎิทิน</td>
-                                        <td className="whitespace-nowrap text-center px-4 py-2 text-blue-700 cursor-pointer"><Link to="/calendarstudy" state={{classroomId:classroom.classId,classroomInfo:classroom}}>ปฎิทินการเรียน</Link></td>
-                                        <td className="whitespace-nowrap text-center px-4 py-2 text-blue-700 cursor-pointer"><Link to="/calendarholiday" state={{classroomId:classroom.classId,classroomInfo:classroom}}>ปฎิทินวันหยุด</Link></td>      
+
+                                        <td className="whitespace-nowrap text-center px-4 py-2 flex gap-2 justify-center">
+                                            <Link to="/calendarstudy" className="cursor-pointer bg-blue-300/60 text-blue-500 px-5 py-[2px] rounded-sm hover:bg-blue-300/100 hover:text-blue-700" state={{classroomId:classroom.classId,classroomInfo:classroom}}>
+                                                ปฎิทินการเรียน
+                                            </Link>
+                                            <Link to="/calendarholiday" className="cursor-pointer bg-blue-300/60 text-blue-500 px-5 py-[2px] rounded-sm hover:bg-blue-300/100 hover:text-blue-700" state={{classroomId:classroom.classId,classroomInfo:classroom}}>
+                                                ปฎิทินวันหยุด
+                                            </Link>
+                                        </td>
+                                    
+                                        <td className="whitespace-nowrap text-center px-4 py-2 text-red-700 cursor-pointer" onClick={() => handelDeleteStudingTime(classroom.classId)}>
+                                            <button className="cursor-pointer bg-red-200 text-red-600 px-5 py-[2px] rounded-sm hover:bg-red-400 hover:text-red-700">
+                                                ลบปฎิทิน
+                                            </button>
+                                            
+                                        </td>
+                                             
                                     </tr>
                                 ))
                             }
                             {
                                 sliceHolidayList.length === 0 && (
                                     <tr>
-                                        <td colSpan={4} className="whitespace-nowrap text-center px-4 py-2 text-gray-700">ไม่มีห้องเรียน</td>
+                                        <td colSpan={5} className="whitespace-nowrap text-center px-4 py-2 text-gray-700">ไม่มีห้องเรียน</td>
                                     </tr>
                                 )
                             }
